+++ example/lib/constants.dart
@@ -1,16 +1,14 @@
 // EXAMPLE VALUES
-const String MNEMONIC =
+const MNEMONIC =
     "web theme zero talk usual whale series giggle crucial quiz crater negative";
-const String PATH = "m/44'/60'/0'/0/0";
-const String ENTITY_ENTRY_POINT = "https://rpc.slock.it/goerli";
-const String ENTITY_ID =
-    "0x62cd5fb8d6874ffc1862ac3384f3149c3a2a62dc5b646cd42d7c0dc26f0e771b";
-const String NETWORK_ID = "goerli";
-const String CENSUS_MERKLE_ROOT =
+const PATH = "m/44'/60'/0'/0/0";
+const ENTITY_ID =
+    "0x8721ba1b56da230be4eb1ce710884024bbb33e073ff39e00ee56b285ba386efa";
+const WEB3_ENTRY_POINT = "https://dai.poa.network";
+const NETWORK_ID = "xdai";
+const CENSUS_MERKLE_ROOT =
     "0x1e3cde94b33218b952d069ef90216661bb244eba3a3a4a774e73952efe8c7110";
-const String CENSUS_PUB_KEY_CLAIM =
-    "Lf+khO3vB+iZ7KHEVfT5lMvYuYa4u9eMejg4iD2jBKM=";
-const String PROCESS_ID =
+const CENSUS_PUB_KEY_CLAIM = "Lf+khO3vB+iZ7KHEVfT5lMvYuYa4u9eMejg4iD2jBKM=";
+const PROCESS_ID =
     "0x239333d368917b24f73940ad17fa5c530a95889895be4a201c46bfc44c787f87";
-const String BOOTNODES_URL_RW =
-    "https://bootnodes.vocdoni.net/gateways.dev.json";
+const BOOTNODES_URL_RW = "https://bootnodes.vocdoni.net/gateways.json";