+++ example/lib/metadata.dart
@@ -28,7 +28,6 @@ class _MetadataScreenState extends State<MetadataScreen> {
     String entityMetaStr;
     EntityReference entity = EntityReference();
     entity.entityId = ENTITY_ID;
-    entity.entryPoints.addAll(["https://rpc.slock.it/goerli"]);
 
     try {
       final gwInfo =
@@ -39,8 +38,9 @@ class _MetadataScreenState extends State<MetadataScreen> {
       final entityMeta = await fetchEntity(entity, dvoteGw, web3Gw);
       entityMetaStr = jsonEncode(entityMeta.toString());
 
-      final String pid = entityMeta.votingProcesses?.active?.first;
-      if (pid is String) {
+      if ((entityMeta.votingProcesses?.active?.length is int) &&
+          entityMeta.votingProcesses.active.length > 0) {
+        final pid = entityMeta.votingProcesses?.active?.first;
         final processMeta = await getProcessMetadata(pid, dvoteGw, web3Gw);
         processMetaStr = processMeta.toString();
         // print(processMetaStr);
@@ -97,7 +97,8 @@ $_processMetaStr''';
             padding: EdgeInsets.all(16),
             child: Column(
               children: <Widget>[
-                Text(entityMeta),Text(processMeta), 
+                Text(entityMeta),
+                Text(processMeta),
               ],
             ),
           ),
@@ -106,25 +107,3 @@ $_processMetaStr''';
     );
   }
 }
-
-// Future<void> metadata() async {
-//   EntityReference entity = EntityReference();
-//   entity.entityId = ENTITY_ID;
-//   entity.entryPoints.addAll(["https://rpc.slock.it/goerli"]);
-
-//   try {
-//     final gwInfo = await getRandomGatewayDetails(BOOTNODES_URL_RW, NETWORK_ID);
-//     final dvoteGw = DVoteGateway(gwInfo.dvote, publicKey: gwInfo.publicKey);
-//     final web3Gw = Web3Gateway(gwInfo.web3);
-
-//     final entityMeta = await fetchEntity(entity, dvoteGw, web3Gw);
-
-//     final String pid = entityMeta.votingProcesses?.active?.first;
-//     if (pid is String) {
-//       final processMeta = await getProcessMetadata(pid, dvoteGw, web3Gw);
-//       print(jsonEncode(processMeta));
-//     }
-//   } catch (err) {
-//     print(err);
-//   }
-// }